refactor(auth): split pages into server components with client form components
